Extract register form header
